+++ docker.sh
@@ -4,9 +4,10 @@ export HOMIE_MQTT_URI="mqtt://synology"
 export HOMIE_MQTT_ID="homie45-bridge-123"
 export HOMIE_DOMAIN4="homie"
 export HOMIE_DOMAIN5="homie"
-export HOMIE_SUBSCRIBE_DELAY=10000
-export HOMIE_LOG_LOGLEVEL="debug"
-# export HOMIE_LOG_LOGLEVEL="info"
+export HOMIE_DEVICE_ID="homie45bridge"
+export HOMIE_SUBSCRIBE_DELAY=1000
+# export HOMIE_LOG_LOGLEVEL="debug"
+export HOMIE_LOG_LOGLEVEL="info"
 
 
 
@@ -18,6 +19,7 @@ docker run -it --rm \
     -e HOMIE_MQTT_ID \
     -e HOMIE_DOMAIN4 \
     -e HOMIE_DOMAIN5 \
+    -e HOMIE_DEVICE_ID \
     -e HOMIE_SUBSCRIBE_DELAY \
     -e HOMIE_LOG_LOGLEVEL \
     tieske/homie45bridge:dev